+++ src/CharacterAtributes/Constitution.java
@@ -1,9 +1,11 @@
 package CharacterAtributes;
 
+import java.math.BigDecimal;
+
 public class Constitution extends Characteristic{
     private String type;
 
-    public Constitution(int value) {
+    public Constitution(BigDecimal value) {
         super(value);
         this.type = "Physical";
     }
