+++ src/CharacterAtributes/Dexterity.java
@@ -1,9 +1,11 @@
 package CharacterAtributes;
 
+import java.math.BigDecimal;
+
 public class Dexterity extends Characteristic{
     private String type;
 
-    public Dexterity(int value) {
+    public Dexterity(BigDecimal value) {
         super(value);
         this.type = "Physical";
     }
