+++ src/CharacterAtributes/Intelligence.java
@@ -1,9 +1,11 @@
 package CharacterAtributes;
 
+import java.math.BigDecimal;
+
 public class Intelligence extends Characteristic{
     private String type;
 
-    public Intelligence(int value) {
+    public Intelligence(BigDecimal value) {
         super(value);
         this.type = "Mental";
     }
