+++ src/CharacterAtributes/Strength.java
@@ -1,9 +1,11 @@
 package CharacterAtributes;
 
+import java.math.BigDecimal;
+
 public class Strength extends Characteristic{
     private String type;
 
-    public Strength(int value) {
+    public Strength(BigDecimal value) {
         super(value);
         this.type = "Physical";
     }
