+++ src/CharacterAtributes/Wisdom.java
@@ -1,9 +1,11 @@
 package CharacterAtributes;
 
+import java.math.BigDecimal;
+
 public class Wisdom extends Characteristic{
     private String type;
 
-    public Wisdom(int value) {
+    public Wisdom(BigDecimal value) {
         super(value);
         this.type = "Mental";
     }
